Moving tokenized card logic to own class (ElevateTokenizedGift)
