+++ src/lwc/geFormRenderer/elevateCaptureGroup.js
@@ -17,15 +17,17 @@ class ElevateCaptureGroup {
         try {
             this._runCount = 1;
 
-            let response = await apexAddToCaptureGroup(tokenizedGift, null);
+            let response = await apexAddToCaptureGroup(tokenizedGift, this.elevateBatchId);
+            console.log(`Add response = ${response}`);
             return Object.assign(new AuthorizedGift, JSON.parse(response).body); 
         } catch (ex) {
             if (!this._runCount) {
                 this.elevateBatchId = await this.create();
                 this.add(tokenizedGift);
+            } else {
+                throw Error; // Propagate error to calling function for error handling
             }
         }
-
     }
 
     async create() {
